// src/types/investmentAmountHistory.ts
export type InvestmentAmountHistoryEntry = {
  amount: string;
  createdAt: number;
  createdDate: string;
  createdTime: string;
  description: string;
  newAmount: string;
  previousAmount: string;
  type: string;
};

export type InvestmentReinvestmentEntry = {
  amount: string;
  createdAt: number;
  createdDate: string;
  createdTime: string;
};